Update adminportal.html
Added spaces next to username and password for user to type in their info. Linked the page when user successfully signs-in so they can continue by changing details in their account, choosing service options, etc. Deleted the part that linked to delete account so user can do so after signing-in.